requirements.txt
absl-py==2.0.0
astunparse==1.6.3
attr==0.3.2
attrs==23.1.0
boto3==1.28.62
botocore==1.31.62
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cloudpickle==2.2.1
configparser==6.0.0
contextlib2==21.6.0
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
Cython==3.0.3
decorator==5.1.1
dm-tree==0.1.8
filelock==3.12.4
flatbuffers==2.0.7
fonttools==4.43.1
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
imageio==2.31.5
Jinja2==3.1.2
jmespath==1.0.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.1
nnabla==1.36.0
nnabla-converter==1.36.0
nnabla-ext-cuda116==1.36.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.12.0
onnx-tf==1.10.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
ply==3.11
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.7.0
scipy==1.11.3
six==1.16.0
sympy==1.12
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.4
tensorflow-addons==0.21.0
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-probability==0.16.0
termcolor==2.3.0
tf2onnx==1.14.0
tflite==2.10.0
tflite2onnx==0.4.1
torch==2.1.0
torchaudio==2.1.0
torchsummary==1.5.1
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Werkzeug==3.0.0
wrapt==1.15.0
